valen: Implement a built-in campaignd probe
The probe deployed is a [request_terms] request, which is acted upon by
all current campaignd versions by sending a [message] WML node
containing a single 'message' attribute containing the server's add-on
licensing terms notice (used by clients as a prompt when uploading
add-ons) as a string.

This is much cheaper than invoking wesnoth_addon_manager -l and waiting
on a humongous add-ons list that we actually ignore every time.