valen: Backend version 2 with new output format and semantics
The old mechanism to define wesnoth.org facilities proved extremely
inflexible and not particularly extensible given that the various
overall facilities (DNS, http, wesnothd, campaignd) had to be handled
directly in the probing section and some of the checks weren't
sufficiently in-depth to be useful (DNS was only done for the
wesnoth.org hostname, for example, and not much information about the
test could be passed to the web front-end).

In order to solve these shortcomings, we are now properly defining a
"facility" concept for each wesnoth.org facility, each one having an
associated IP address, description, protocol, and other information.
This should make it far easier to update existing entries or even add
new ones in the future (or even read this configuration from a separate
file!).

More information is now recorded for the front-end for every single
facility, and the output file is now in JSON format so that we can do
whatever we want with the report without worrying about its structure.

The front-end still needs to be updated to understand JSON and the new
facility-based reporting, though.