valen: Add response timing info in milliseconds to probe reports
It might be worth it to add timing reports to the front-end at a later
point. In order to enable this, we have to keep the temporary otimer
objects formerly created by check_url() and check_wesnothd() in a scope
that can be reached by the report generation code, hence this patch is a
bit more noisy than expected.