web: Use a minimal version of Glamdrol for a base CSS stylesheet
Compared to glamdrol/main.css which has 1184 lines and is 16.5 KiB in
size, glamdrol/mini.css has 329 lines in 4.3 KiB. It only contains the
styles I deemed relevant for generic pages, without all the
MediaWiki-sepcific cruft normally needed for the wesnoth.org proper.

With this change, the original big Glamdrol stylesheet is no longer
required, so it's now gone.