valen: Stop monitoring basilic (server2.w.o)
The provider, tuxfamily, is winding down and as a result, server2.w.o
has remained unavailable for multiple months in a row.

<https://forum.tuxfamily.org/post/3381/#p3381>

There is currently no intention to provide an official alternative
server controlled by wesnoth.org staff, so for the foreseeable future
it's just going to be the primary MP server.