valen: Increase wml and http clients' timeout from 10s to 25s
Unfortunately, it is considered Normal(tm) for wesnoth.org facilities
to take an excessing amount of time to respond. Ideally valen2 should
report timings and warn about them without deeming the affected
facility to be offline, but we're not there yet.

We could increase this from 10s to 60s, but it would have the
unfortunate consequence of causing valen to take too long to report
back, since we currently perform requests sequentially and not in
separate subprocesses or threads. Again, this is something that needs
to be taken care of by valen2 instead of trying to fix it in this old
crusty codebase.