wesmere: Move image resources to img/
For some reason, it escaped my notice all these years that the logo
references in mini.css point to a file that does not exist on the live
version at the location it wants. While a possible fix for this is to
update mini.css to point to the image files in wesmere/, this is more
trouble than it's worth since it requires additional CSS/SCSS changes
to mini.css, which would make the code look a little awkward. Namely,
it would necessitate something like this at the end of mini.scss in
the Wesmere source:

  #logo > a {
    @include fixed-size-high-dpi-image(
      url('logo-minimal-64.png'),
      url('logo-minimal-64@2x.png'),
      64, 64);
  }

(Admittedly, I kinda like this better than the current version, but I'm
really trying to avoid changing Wesmere for the sake of Valen because
it means breaking any references to these paths that there may be
lying around in places with old copies of Wesmere, etc.)

Instead, just move the files to img/ so we can appease the stylesheet
without any modifications, and also throw the favicons in there while
we are at it.